packages/backend/requirements.txt
flask
google-cloud-bigquery
openai=1.2.0
tiktoken
tenacity
requests
termcolor
python-dotenv
waitress
gunicorn